% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/INLA_Approximation.R
\name{calculate.moller.hetero2}
\alias{calculate.moller.hetero2}
\title{calculate.moller.hetero}
\usage{
calculate.moller.hetero2(tree, lengthout, prec_alpha = 0.01,
  prec_beta = 0.01, E.log.zero = -100, alpha = NULL, beta = NULL)
}
\arguments{
\item{lengthout}{number of grid points}

\item{prec_alpha}{alpha gamma hyperparameter of precision parameter for GP prior}

\item{prec_beta}{beta gamma hyperparameter of precision parameter for GP prior}

\item{E.log.zero}{internal}

\item{alpha}{TODO}

\item{beta}{TODO}

\item{coal.factor}{is a vector with coalescent times in increasing order}

\item{s}{sampling times and coalescent times in increasing order}

\item{event}{and indicator vector with 1 for coalescent times and 0 for sampling times that correspond to the s vector}
}
\description{
Approximates the posterior distribution of Ne from a single genealogy at a regular grid of points using INLA package
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

