% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/INLA_Approximation.R
\name{calculate.moller}
\alias{calculate.moller}
\title{calculate.moller}
\usage{
calculate.moller(tree, lengthout, L = 1)
}
\arguments{
\item{lengthout}{number of grid points}

\item{L}{the length for the definition of the grid}

\item{data1}{is a dataframe with two columns. The first column has the intercoalescent times and the second column the number of lineages}
}
\description{
Approximates the posterior distribution of Ne from a single genealogy at a regular grid of points using INLA package
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

