% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{conf.int.skyline}
\alias{conf.int.skyline}
\title{conf.int.skyline}
\usage{
conf.int.skyline(conf.int, epsilon = 0, plot_type = c("step"), ...)
}
\arguments{
\item{conf.int:}{this type of output from Phylos2Skylines_anchor}

\item{epsilon:}{a smoothing parameter. Any time intervals smaller than this value will be joined with adjacent bins until all are wider than this.}

\item{plot_type:}{choose between "step", where adjacent points on the skyline are joined by a vertical and a horizontal line, and "linear", where adjacent points on the skyline are joined by a single sloping straight line.}
}
\value{
A plot with median in black and confidence intervals in red
}
\description{
Makes a skyline plot from a matrix with medians and confidence intervals as produced by Phylos2Skylines_anchor
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

