% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{draw_skyline}
\alias{draw_skyline}
\title{draw_skyline}
\usage{
draw_skyline(end_times, pop_size, plot_type = c("linear", "step"), ...)
}
\value{
Draws the lines of a skyline (step) or INLA (piece-wise linear)
}
\description{
Draws a classic skyline onto a plot
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

