% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequencewrapper.R
\name{eq.prob}
\alias{eq.prob}
\title{eq.prob}
\usage{
eq.prob(pi = NULL, n.prob = NULL)
}
\value{
A vector of pi values.
}
\description{
Checks that appropriate equilibrium probabilities are specified, otherwise specifies equal probabilities.
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

