% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trees.R
\name{gen_INLA_args}
\alias{gen_INLA_args}
\title{gen_INLA_args}
\usage{
gen_INLA_args(coal_times, s_times, n_sampled)
}
\arguments{
\item{coal_times}{indicates the coalescent times}

\item{s_times}{indicates the sampling times}

\item{n_sampled}{the number of samples}
}
\description{
Used for heterochronous sampling. It has nothing to do with INLA
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

