% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNe.R
\docType{data}
\name{genObj}
\alias{genObj}
\title{varNe_point}
\format{An object of class \code{character} of length 1.}
\usage{
genObj
}
\arguments{
\item{genObj}{A genind object with years as populations.}
}
\value{
a vector with
}
\description{
Calculates population size from a genind object by calculating the variance of allele frequencies between generations.

The variance F is inversely proportional to the effective population size Ne.

This function calculates F based on Jorde and Ryman 2007. They state hat low frequency alleles are not problematic / do not bias Ne estimates, thus they are not removed here.
}
\examples{
library(adegenet)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}
\keyword{datasets}

