% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{getNodeAges}
\alias{getNodeAges}
\title{getNodeAges}
\usage{
getNodeAges(x, from_past = F)
}
\arguments{
\item{x}{A phylo object}

\item{from_past}{If FALSE (default) the node ages increase backwards in time, with the most recent tip being zero. If TRUE the root node is zero and the node ages increase towards the tips of the tree.}
}
\value{
For a tree with n tips, the vector is <tip1, tip2...tipn, root, internal_node2, internal_node3>
}
\description{
Returns a vector with the depths, or heights of the nodes in the tree
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

