% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HENe.R
\name{getSampleSize_byLocus}
\alias{getSampleSize_byLocus}
\title{getSampleSize_byLocus}
\usage{
getSampleSize_byLocus(genind_obj)
}
\description{
Takes a genind object and returns the number of samples occurring at each locus
}
\examples{
library(adegenet)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

