% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HENe.R
\name{HENe}
\alias{HENe}
\title{HENe}
\usage{
HENe(genind_obj)
}
\description{
Calculates population size from a genind object by heterozygote excess
}
\examples{
library(adegenet)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>, George Shirreff <georgeshirreff@gmail.com>
}

