% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simulate_trees.R
\name{heterochronous.gp.stat}
\alias{heterochronous.gp.stat}
\title{heterochronous.gp.stat}
\usage{
heterochronous.gp.stat(phy)
}
\arguments{
\item{phy}{takes a phylo object}
}
\description{
Generates sufficient statistcs
}
\author{
Julia Palacios and Vladimir Minin \email{julia.pal.r@gmail.com}
}

