% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNe.R
\name{jack_samples}
\alias{jack_samples}
\title{jackknifing samples}
\usage{
jack_samples(genObj)
}
\arguments{
\item{genObj}{A genind object with years as populations.}
}
\value{
a list. Each element is a jackknifed genind object
}
\description{
jackknifes samples
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

