% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDNe.R
\name{LDNe_point}
\alias{LDNe_point}
\title{Calculating point estimate of linkage disequilibrium effective population size}
\usage{
LDNe_point(genObj, crit = 0.01, mating = "random")
}
\arguments{
\item{genObj}{genind object}

\item{crit}{the lowest allele frequency used in the calculation. Alleles with lower allele frequency are excluded}

\item{mating}{mating system information}
}
\description{
Takes a genind object and calculates effective population size based on the magnitude of linkage disequilibrium found in a sample
}
\examples{
library(adegenet)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

