% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDNe.R
\name{makeCrit}
\alias{makeCrit}
\title{Removing alleles below critical frequency}
\usage{
makeCrit(genObj, crit = 0.01)
}
\arguments{
\item{genObj}{genind object}

\item{crit}{the lowest allele frequency allowed to occur. Alleles at a lower frequency are excluded.}
}
\description{
Takes a genind object, removes alleles below a certain allele frequency, and returns a list of genind objects, one object per locus
}
\examples{
library(adegenet)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

