% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDNe.R
\name{makeP}
\alias{makeP}
\title{Allele frequencies per locus}
\usage{
makeP(genind.by.locus)
}
\arguments{
\item{genind.by.locus}{list of genind objects, one list element per locus}
}
\description{
Takes a genind object and calculates the frequency of each allele per locus
}
\examples{
library(adegenet)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

