% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ms2ms.R
\name{ms2ms}
\alias{ms2ms}
\title{ms2ms}
\usage{
ms2ms(ms.output)
}
\arguments{
\item{ms.output}{Output from ms (library phyclust)}
}
\value{
A genind object of simulated microsatellite data.
}
\description{
Simulates microsatellite data by implementing the code of Pidugu and Schlotterer (2006) in R.
}
\examples{
library(phyclust)
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

