% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequencewrapper.R
\name{ms2nuc}
\alias{ms2nuc}
\title{ms2nuc}
\usage{
ms2nuc(ms.res, fileout)
}
\value{
A table.
}
\description{
Converts output of ms2ms to nucleotides.
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

