% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequencewrapper.R
\name{phang2nuc}
\alias{phang2nuc}
\title{phang2nuc}
\usage{
phang2nuc(sq = phang.seq, file = "test.out")
}
\value{
No output.
}
\description{
Converts output of phangorn to nucleotides.
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

