% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequencewrapper.R
\name{phylosim2nuc}
\alias{phylosim2nuc}
\title{phylosim2nuc}
\usage{
phylosim2nuc(align = sim, file = "test.out")
}
\value{
A table.
}
\description{
Converts PhyloSim output to phylip format.
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

