% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/INLA_Approximation.R
\name{plot_INLA}
\alias{plot_INLA}
\title{plot_INLA}
\usage{
plot_INLA(INLA_out, traj = NULL, xlim = NULL, ...)
}
\arguments{
\item{INLA_out}{Otput from the inla functions}

\item{traj}{the true trajectory}

\item{xlim}{}
}
\description{
Plots the output from the inla functions for Ne
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

