% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/INLA_Approximation.R
\name{plot_INLA2}
\alias{plot_INLA2}
\title{plot_INLA2}
\usage{
plot_INLA2(result, traj = NULL, xlim = NULL, ...)
}
\arguments{
\item{result}{Matrix with four columns}

\item{traj}{the true trajectory}

\item{xlim}{(optional)}
}
\description{
Plots a general output with 4 columns: time, median, lower and upper bound
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

