% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INLA_Approximation.R
\name{plot_INLA3}
\alias{plot_INLA3}
\title{plot_INLA3}
\usage{
plot_INLA3(result, traj = NULL, xlim = NULL, ...)
}
\arguments{
\item{result}{Matrix with many columns}

\item{traj}{the true trajectory}

\item{xlim}{(optional)}
}
\description{
Plots bunch of trees
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

