% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequencewrapper.R
\name{sequence.wrapper}
\alias{sequence.wrapper}
\title{sequence.wrapper}
\usage{
sequence.wrapper(n = 10, L = 10, method = "", ms.args = NULL,
  tree = NULL, pi.n = 0.25, kappa = "", outfile = "out",
  ancseq = NULL, rate.in = 1)
}
\arguments{
\item{n}{number of sequences}

\item{L}{length of sequence}

\item{method}{choice of sequence simulation}

\item{ms.args}{arguments passed to ms}

\item{tree}{tree of format phylo}

\item{outfile}{name of the file that writes out}
}
\value{
Writes out sequences in phylip format to your working directory.
}
\description{
A wrapper function for the sequence simulator programs ms, seqgen, simSeq, and PhyloSim.
}
\examples{
library(phyclust)
}
\author{
Rebecca Harris <rbharris@uw.edu>
}

