% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trees.R
\name{simulate.tree}
\alias{simulate.tree}
\title{simulate.tree}
\usage{
\method{simulate}{tree}(n = 10, N = 1, sampling = "iso",
  args = "-T -G 0.1", Ne = 1, max = 1, simulator = NULL,
  sample = NULL, ...)
}
\arguments{
\item{n}{Number of samples}

\item{N}{Number of genealogies or trees}

\item{sampling}{It can take two values: iso or hetero. If the sampling scheme is (iso) it generates an isochronous genealogy (all samples at time 0). Otherwise, it assumes it is heterochronous sampling.If hetero, you need to specify samples}

\item{args}{String of characters for ms}

\item{Ne}{The standard is set to 1. It takes any function as input for the demographic model}

\item{max}{(optional) for thinning simulator. It is an upper bound on 1/Ne}

\item{simulator}{It can either be null, ms or thinning. If null, it uses ape rcoal.}

\item{sample}{It takes a matrix of two columns. The first column inciates the number of samples and the second column the sampling times}
}
\description{
Simulates genealogies under the coalescent
}
\author{
Julia Palacios \email{julia.pal.r@gmail.com}
}

