% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SkylineDatedTips.R
\name{skyline.datedPhylo}
\alias{skyline.datedPhylo}
\title{skyline.datedPhylo}
\usage{
\method{skyline}{datedPhylo}(tr)
}
\arguments{
\item{An}{object of class "datedPhylo"}
}
\value{
An object of class "skyline"
}
\description{
Like skyline() from ape but takes the class "datedPhylo"
}
\examples{
library(ape)
}
\author{
Lucy Mengqi Li <mengqi.li09@imperial.ac.uk>
}

