% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineDatedTips.R
\name{skyline.with.sampling}
\alias{skyline.with.sampling}
\title{skyline.with.sampling}
\usage{
skyline.with.sampling(ci)
}
\arguments{
\item{ci:}{a datedCI object}
}
\value{
An object of class datedSkyline
}
\description{
For a given datedCI object return the skyline for each coalescent interval
}
\examples{
library(ape)
}
\author{
Lucy Mengqi Li <mengqi.li09@imperial.ac.uk>
}

