% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{skyline_smooth}
\alias{skyline_smooth}
\title{skyline_smooth}
\usage{
skyline_smooth(end_times, pop_size, epsilon = 0, plot_type = c("step"),
  draw = T, ...)
}
\value{
Draws the lines of a skyline, but with a minimum time interval specified by epsilon
}
\description{
Draws a generalized skyline onto a plot
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

