% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNe.R
\name{sum_jack2}
\alias{sum_jack2}
\title{jackknifing samples}
\usage{
sum_jack2(Ne_point, bs, statistic)
}
\arguments{
\item{Ne_point}{output of function varNe_point. Matrix containing point estimates of Ne, one estimate per comparison}

\item{bs}{jackknifed populations (list of genind objects)}

\item{statistic}{function to be employed, here varNe_point}
}
\value{
a list. Each element is a jackknifed genind object
}
\description{
jackknifes samples
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

