% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarNe.R
\name{varNe}
\alias{varNe}
\title{varNe}
\usage{
varNe(genObj)
}
\arguments{
\item{genObj}{A genind object with years as populations.}
}
\value{
a data.frame. Each column reports the result for a comparison of two generations,

and the rows correspond to the generations compared, the point estimate of variance Ne, and the summary statistics of the jackknifing (median, 0.025 quantile, 0.975 quantile, variance)
}
\description{
Calculates effective population size from a genind object by calculating the variance of allele frequencies between generations, and jackknifes the samples.

The variance F is inversely proportional to the effective population size Ne.

This function calculates F based on Jorde and Ryman 2007. They state hat low frequency alleles are not problematic / do not bias Ne estimates, thus they are not removed here.
}
\examples{
data(simG3)
}
\author{
Christine Ewers-Saucedo <ewers.christine@gmail.com>
}

