% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineDatedTips.R
\name{coalescent.intervals.datedPhylo}
\alias{coalescent.intervals.datedPhylo}
\title{coalescent.intervals.datedPhylo}
\usage{
\method{coalescent.intervals}{datedPhylo}(tr)
}
\arguments{
\item{tr:}{a phylogeny of class "datedPhylo"}
}
\value{
A datedCI object with coalescent intervals in units of substitutions.
}
\description{
Functions to find the skyline of a binary, rooted tree with heterochronous

tips, i.e. sequences were sampled sequentially in time.

The likelihood function is based on Equations 3 and 4 in

Drummond et al (2005) Mol. Biol. Evol.
}
\examples{
library(ape)
}
\author{
Lucy Mengqi Li <mengqi.li09@imperial.ac.uk>
}

