% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{conf.int.skyline}
\alias{conf.int.skyline}
\title{conf.int.skyline}
\usage{
conf.int.skyline(conf.int, epsilon = 0, plot_type = c("step"), ...)
}
\value{
A plot with median in black and confidence intervals in red
}
\description{
Makes a skyline plot from a matrix with medians and confidence intervals as produced by Phylos2Skylines_anchor
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

