% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{draw_skyline}
\alias{draw_skyline}
\title{draw_skyline}
\usage{
draw_skyline(end_times, pop_size, plot_type = c("linear", "step"), col = 1)
}
\arguments{
\item{end_times:}{the times at the end of every interval in the skyline.}

\item{pop_size:}{the effective population size at that time point.}

\item{plot_type:}{choose between "step", where adjacent points on the skyline are joined by a vertical and a horizontal line, and "linear", where adjacent points on the skyline are joined by a single sloping straight line.}
}
\value{
Draws the lines of a skyline (step) or INLA (piece-wise linear)
}
\description{
Draws a classic skyline onto a plot
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

