% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkylineDatedTips_extension.R
\name{skyline_smooth}
\alias{skyline_smooth}
\title{skyline_smooth}
\usage{
skyline_smooth(end_times, pop_size, epsilon = 0, plot_type = c("step"), ...)
}
\arguments{
\item{epsilon:}{a smoothing parameter. Any time intervals smaller than this value will be joined with adjacent bins until all are wider than this.}

\item{draw:}{logical specifying if the skyline should be drawn (TRUE) or simply the points on the graph returned (FALSE)}
}
\value{
Draws the lines of a skyline, but with a minimum time interval specified by epsilon

If draw=F, it returns a list with the end times of the intervals and the population sizes, smoothed as per the value of epsilon.
}
\description{
Draws a generalized skyline onto a plot
}
\examples{
library(ape)
}
\author{
George Shirreff <georgeshirreff@gmail.com>
}

